requirements.txt
flask==2.3.3
flask-cors==4.0.0
pillow==11.1.0
python-dotenv==1.0.0
supabase==2.13.0
requests==2.32.3
modal==0.73.91
litellm==1.63.6
boto3==1.37.11
google-genai==1.7.0
gunicorn==21.2.0
